finalization

+++ app/src/main/java/com/example/wms_scan/data/response/GetCartonDetailsResponse.kt
@@ -18,7 +18,7 @@ class GetCartonDetailsResponse {
 
     @SerializedName("Mat_Stock")
     @Expose
-     val matStock: Int? = null
+     val matStock: Double? = null
 
     @SerializedName("IsExist")
     @Expose
